addresses some of the PR comments
